+++ src/robin_hood.h
@@ -35,6 +35,7 @@
 //
 // - iterators with offsets
 // - num steps an iterator took
+// - better error report
 #ifndef ROBIN_HOOD_H_INCLUDED
 #define ROBIN_HOOD_H_INCLUDED
 
@@ -51,6 +52,8 @@
 #include <string>
 #include <type_traits>
 #include <utility>
+
+#include "utils.hpp"
 
 // #define ROBIN_HOOD_LOG_ENABLED
 #ifdef ROBIN_HOOD_LOG_ENABLED
@@ -1288,9 +1291,7 @@ private:
     size_t insert_move(Node&& keyval) {
         // we don't retry, fail if overflowing
         // don't need to check max num elements
-        if (0 == mMaxNumElementsAllowed && !try_increase_info()) {
-            throwOverflowError();
-        }
+        OVERFLOW(0 == mMaxNumElementsAllowed && !try_increase_info(), "insertion fail");
 
         size_t idx;
         InfoType info;
@@ -1757,9 +1758,7 @@ public:
         while (calcMaxNumElementsAllowed(newSize) < minElementsAllowed && newSize != 0) {
             newSize *= 2;
         }
-        if (ROBIN_HOOD_UNLIKELY(newSize == 0)) {
-            throwOverflowError();
-        }
+        OVERFLOW(ROBIN_HOOD_UNLIKELY(newSize == 0), "reserve fail.");
 
         rehashPowerOfTwo(newSize);
     }
@@ -1827,9 +1826,7 @@ public:
         auto const total64 = ne * s + infos;
         auto const total = static_cast<size_t>(total64);
 
-        if (ROBIN_HOOD_UNLIKELY(static_cast<uint64_t>(total) != total64)) {
-            throwOverflowError();
-        }
+        OVERFLOW(ROBIN_HOOD_UNLIKELY(static_cast<uint64_t>(total) != total64), "calcNumBytesTotal fail");
         return total;
 #endif
     }
@@ -2041,9 +2038,7 @@ private:
                                        << (static_cast<double>(mNumElements) * 100.0 /
                                            (static_cast<double>(mMask) + 1)));
         // it seems we have a really bad hash function! don't try to resize again
-        if (mNumElements * 2 < calcMaxNumElementsAllowed(mMask + 1)) {
-            throwOverflowError();
-        }
+        OVERFLOW(mNumElements * 2 < calcMaxNumElementsAllowed(mMask + 1), "increase_size fail");
 
         rehashPowerOfTwo((mMask + 1) * 2);
     }