+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Update multigpu documentation
 - Update GIL release
 - Minor error fixes on `examples/modelnet40.py`
+- CoordsMap size initialization updates
 
 
 ## [0.3.1] - 2019-12-15